Major: Added entity loading to map files (TODO: replace switch statement for entity type with some other more modular system, so that FileHandler does not need to be edited as new entity types are added) Added a second prefab entity: Snail (the snail chases the player)
Minor:
Cleaned up excessive try/catch blocks in file loading
Minor polish throughout

+++ src/main/RNG.java
@@ -12,12 +12,12 @@ public class RNG {
 
     //upTo(): returns a floating point number from 0 to maximum
     public static float upTo(int maximum){
-        return (float)ThreadLocalRandom.current().nextDouble(maximum);
+        return ThreadLocalRandom.current().nextFloat(maximum);
     }
 
     //maxDistance(): returns a floating point number from negative maximum to positive maximum
     public static float maxDistance(int maximum){
-        return (float)ThreadLocalRandom.current().nextDouble(-maximum, maximum);
+        return ThreadLocalRandom.current().nextFloat(-maximum, maximum);
     }
 
     //TODO: MORE RNG FUNCTIONS. AS I NEED THEM.
